requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.7
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
nltk==3.4
numpy==1.15.4
oauthlib==2.1.0
protobuf==3.6.1
PySocks==1.6.8
pytz==2018.7
PyYAML>=4.2b1
requests==2.20.1
requests-oauthlib==1.0.0
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
textblob==0.15.2
tweepy==3.7.0
urllib3==1.24.1
Werkzeug==0.14.1
